extract NamedParamMetas
