+++ src/NamedParamMetasInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of the BEAR.Resource package.
+ *
+ * @license http://opensource.org/licenses/MIT MIT
+ */
+namespace BEAR\Resource;
+
+interface NamedParamMetasInterface
+{
+    public function __invoke(callable $callable) : array;
+}
